Refactor methods from Tokenizer to other classes like StringUtil and TodoList
