+++ src/main/java/seedu/toluist/commons/util/StringUtil.java
@@ -2,6 +2,9 @@ package seedu.toluist.commons.util;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * Helper functions for handling strings.
@@ -54,5 +57,75 @@ public class StringUtil {
      */
     public static boolean isUnsignedInteger(String s) {
         return s != null && s.matches("^0*[1-9]\\d*$");
+    }
+
+    /**
+     * Splits a string of indexes into a list.
+     *   Examples:
+     *   splitIndexes(" - 3",           8) -> [1, 2, 3]
+     *   splitIndexes(" 3 -",           8) -> [3, 4, 5, 6, 7, 8]
+     *   splitIndexes("3 4,  5",        8) -> [3, 4, 5]
+     *   splitIndexes("3 - 5",          8) -> [3, 4, 5]
+     *   splitIndexes("- 3, 5, 7 - 12", 8) -> [1, 2, 3, 5, 7, 8]
+     * This function is robust enough to handle excessive white spaces (" ") and commas (",").
+     * @param stringIndexes, the unprocessed string of indexes from user's input
+     * @param maxIndex, the maximum possible index number
+     * @return a list of integer containing all valid indexes
+     */
+    public static List<Integer> splitIndexes(String stringIndexes, int maxIndex) {
+        // Prepare stringIndexes in the correct format to be processed
+        // Correct format example: ["2", "-", "5", "7", "11", "-", "13", "15"]
+        String processedStringIndexes = stringIndexes.replaceAll("-", " - ");
+        String[] splittedStringIndexes = processedStringIndexes.split(" |\\,");
+        splittedStringIndexes = Arrays.
+                stream(splittedStringIndexes).
+                filter(s -> !s.isEmpty()).
+                toArray(String[]::new);
+        for (String splittedStringIndex: splittedStringIndexes) {
+            System.out.print(" " + splittedStringIndex);
+        }
+        System.out.println("");
+
+        // Process formatted stringIndexes
+        List<Integer> indexes = new ArrayList<Integer>();
+        int i = 0;
+        while (i < splittedStringIndexes.length) {
+            String splittedStringIndex = splittedStringIndexes[i];
+
+            if (StringUtil.isUnsignedInteger(splittedStringIndex)) {
+
+                int index = Integer.valueOf(splittedStringIndex);
+                if (index <= maxIndex) {
+                    indexes.add(Integer.valueOf(splittedStringIndex));
+                    i += 1;
+                } else {
+                    // Invalid state, early termination
+                    return indexes;
+                }
+
+            } else if (splittedStringIndex.equals("-")) {
+
+                // If stringIndexes starts with "-", the startIndex will be 0;
+                int startIndex = (indexes.isEmpty()) ? 0 : indexes.get(indexes.size() - 1);
+                // If stringIndexes ends with "-", the endIndex will be todoListSize
+                int endIndex = maxIndex;
+                if (i + 1 < splittedStringIndexes.length
+                        && StringUtil.isUnsignedInteger(splittedStringIndexes[i + 1])) {
+                    endIndex = Integer.valueOf(splittedStringIndexes[i + 1]);
+                } else if (i + 1 > splittedStringIndexes.length) {
+                    // Invalid state, early termination
+                    return indexes;
+                }
+                for (int value = startIndex + 1; value <= Integer.min(endIndex, maxIndex); value++) {
+                    indexes.add(value);
+                }
+                i += 2;
+
+            } else {
+                // Invalid state, early termination
+                return indexes;
+            }
+        }
+        return indexes;
     }
 }
