+++ src/main/java/seedu/toluist/controller/commons/TaskTokenizer.java
@@ -1,15 +1,8 @@
 package seedu.toluist.controller.commons;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import seedu.toluist.commons.util.StringUtil;
-import seedu.toluist.model.Task;
-import seedu.toluist.ui.UiStore;
 
 /**
  * Tokenize task into index, description, start date and/or end date.
@@ -21,7 +14,6 @@ public class TaskTokenizer {
     public static final int INVALID_INDEX = -1;
 
     public static final String TASK_VIEW_INDEX = "index";
-    public static final String TASK_VIEW_INDEXES = "indexes";
     public static final String TASK_DESCRIPTION = "description";
     public static final String TASK_START_DATE_KEYWORD = "startdate/";
     public static final String TASK_END_DATE_KEYWORD = "enddate/";
@@ -62,78 +54,5 @@ public class TaskTokenizer {
             tokens.put(TASK_END_DATE_KEYWORD, endDate);
         }
         return tokens;
-    }
-
-    public List<Integer> splitIndexes(String stringIndexes, int todoListSize) {
-        // Prepare stringIndexes in the correct format to be processed
-        // Correct format example: ["2", "-", "5", "7", "11", "-", "13", "15"]
-        String processedStringIndexes = stringIndexes.replaceAll("-", " - ");
-        String[] splittedStringIndexes = processedStringIndexes.split(" |\\,");
-        splittedStringIndexes = Arrays.
-                stream(splittedStringIndexes).
-                filter(s -> !s.isEmpty()).
-                toArray(String[]::new);
-        for (String splittedStringIndex: splittedStringIndexes) {
-            System.out.print(" " + splittedStringIndex);
-        }
-        System.out.println("");
-
-        // Process formatted stringIndexes
-        List<Integer> indexes = new ArrayList<Integer>();
-        int i = 0;
-        while (i < splittedStringIndexes.length) {
-            String splittedStringIndex = splittedStringIndexes[i];
-
-            if (StringUtil.isUnsignedInteger(splittedStringIndex)) {
-
-                int index = Integer.valueOf(splittedStringIndex);
-                if (index <= todoListSize) {
-                    indexes.add(Integer.valueOf(splittedStringIndex));
-                    i += 1;
-                } else {
-                    // Invalid state, early termination
-                    return indexes;
-                }
-
-            } else if (splittedStringIndex.equals("-")) {
-
-                // If stringIndexes starts with "-", the startIndex will be 0;
-                int startIndex = (indexes.isEmpty()) ? 0 : indexes.get(indexes.size() - 1);
-                // If stringIndexes ends with "-", the endIndex will be todoListSize
-                int endIndex = todoListSize;
-                if (i + 1 < splittedStringIndexes.length
-                        && StringUtil.isUnsignedInteger(splittedStringIndexes[i + 1])) {
-                    endIndex = Integer.valueOf(splittedStringIndexes[i + 1]);
-                } else if (i + 1 > splittedStringIndexes.length) {
-                    // Invalid state, early termination
-                    return indexes;
-                }
-                for (int value = startIndex + 1; value <= Integer.min(endIndex, todoListSize); value++) {
-                    indexes.add(value);
-                }
-                i += 2;
-
-            } else {
-                // Invalid state, early termination
-                return indexes;
-            }
-        }
-        return indexes;
-    }
-
-    public List<Task> getTasks(List<Integer> indexes) {
-        List<Task> tasks = new ArrayList<Task>();
-        for (int index: indexes) {
-            System.out.print(" " + index);
-            tasks.add(UiStore.getInstance().getTasks().get(index - 1));
-        }
-        System.out.println("");
-        return tasks;
-    }
-
-    public Task getTask(String indexToken) {
-        int index = indexToken == null ? INVALID_INDEX : Integer.parseInt(indexToken) - 1;
-        Task task = indexToken == null ? null : UiStore.getInstance().getTasks().get(index);
-        return task;
     }
 }
